Add Entity.java and Player.java
Entity class is abstract and will be inherited by
sub entities in the game. The first such use is the
player entity. Our test square and movement control
is migrated to these classes.

I'm not crazy about the idea of passing the
keyhandler into the Player but I don't yet know of
a better implementation.